[Package] refactor module
